Change current_user assignment in session to user_id
If we assign the whole user, then the whole user is stored in the
cookie. Apart from not needed data in the cookie, this might lead to
using the user from the cookie as initial data when mounting LiveViews.

This is critical, if a LiveView displays user data, because this data
will only change after login and logout.

Also it's cleaner anyway to just pass the user_id through the session to
the LiveViews and reload the user when needed from the database.