Handle revoked api access by logging out user
Users can revoke their OAuth app access on GitHub. An open page would
then try to fetch GitHub resources and would return with a 401 status,
which leads to a 500 page in the app.

So, to prevent this case the user is now just logged out, when api
requests do not return with 200 status codes.